# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biblioteca API - Enterprise Edition",
        "description": "Documentación completa y profesional de la API para el sistema de gestión de biblioteca. Incluye módulos de administración avanzada, trazabilidad (Audit Logs), reportes dinámicos y herramientas operativas (QR, Bulk Import).",
        "contact": {
            "email": "soporte@bibliotecaapi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de API Principal"
        }
    ],
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "List all books",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "isbn",
                                    "year",
                                    "author_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "El señor de los anillos"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "9788478884453"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 1954
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "cover_image": {
                                        "description": "Cover image file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "synopsis": {
                                        "type": "string",
                                        "example": "Una historia fantástica..."
                                    },
                                    "pages": {
                                        "type": "integer",
                                        "example": 423
                                    },
                                    "publisher": {
                                        "type": "string",
                                        "example": "Editorial XYZ"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/books/{book}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a book by ID",
                "operationId": "d8bbccd69b954f8a92bd6c3292b15842",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "operationId": "cf6535a3185059974fb8baaf90b29eaf",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "El señor de los anillos"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "9788478884453"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 1954
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "cover_image": {
                                        "description": "Cover image file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "synopsis": {
                                        "type": "string",
                                        "example": "Una historia fantástica..."
                                    },
                                    "pages": {
                                        "type": "integer",
                                        "example": 423
                                    },
                                    "publisher": {
                                        "type": "string",
                                        "example": "Editorial XYZ"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "operationId": "59b4f85b4a4fe1e69af7004dce3d489f",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "required": [
                    "id",
                    "title",
                    "isbn",
                    "year",
                    "author_id",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Cien años de soledad"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "9780307474728"
                    },
                    "year": {
                        "type": "integer",
                        "example": 1967
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "cover_image_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://api.test/storage/covers/1.jpg"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 15
                    },
                    "synopsis": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}